pyproject.toml
[tool.poetry]
name = "frozen-automata"
version = "0.1.2"
description = ""
authors = ["0xbeef1eaf <151461594+0xbeef1eaf@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
Pillow = "^10.1.0"
pystray = "^0.19.5"
platformdirs = "^4.0.0"
screeninfo = "^0.8.1"
keyboard = "^0.13.5"
pydantic = "^2.5.2"
requests = "^2.31.0"
levenshtein = "^0.23.0"
easyconfig = "^0.3.1"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
dataclasses-json = "^0.6.3"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pyright = "^1.1.337"
pyinstaller = "^6.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
include = ["src"]
exclude = ["**/node_modules",
    "**/__pycache__",
    "src/experimental",
    "src/typestubs"
]
ignore = []
defineConstant = { DEBUG = true }
stubPath = "src/stubs"

reportMissingImports = true
reportMissingTypeStubs = false

pythonVersion = "3.10"
pythonPlatform = "Windows"

executionEnvironments = [
  { root = "src" }
]